feat: handle the client transform manually rather than relying on world
this is a floating point mitigation feature. By parenting on the world, we were essentially getting the engine to do the work of moving an object all they to origin and back again and then add the client transform on top. This is seemingly done in-engine as a single piece of arithmetic (which, why wouldn't it be?).

This fix separates out the two and no longer relies on the engine transform tree - that is to say, it sends the local origin offset along with the local position and each client resolves the remote origin against its' own local where and precision loss may occur. It _then_, after this calculation, adds the position offset with maximum precision.